Improvements and fixes to Game and TicTacToeGame design.
* The diagram for Game was incorrect because it showed an Interface
containing a reference to a GameType, whereas only a Class instance can
do that so the association had to be moved to a concrete class.
- Some pointless methods were removed from TicTacToeGame.
* Some game logic methods which were assumed to be okay as static
methods may actually need some instance variables of the game, not just
the board, such as the rules to decide if a particular game state is a
finish state.