Add more tests of more consecutive moves.
This has highlighted that the full state of the game is in the
TicTacToeGame class, not just the TicTacToeBoard. e.g. which player goes
next. This is an issue because check(..) cannot check if a move is valid
for a player to make on that turn, it can only check it against the
contents of the board. 
Check() is supposed to be a place where all logic of validating a move
is implemented, but to achieve this may require some change to the
object design.