Partially implement playMove, rename isFinal.
The isFinal has been renamed and the return type changed to give more
useful information and preserve value of some of the work that goes into
deciding if the game is over. i.e. it is important to know how the game
ended, not just that it is at an end.